backend/requirements.txt
# --- Core ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.6.4
python-dotenv==1.0.1

google-generativeai==0.3.2

# --- Security & Validation ---
bandit==1.7.8
pip-audit==2.7.1

# --- Testing ---
pytest==8.2.2

streamlit==1.35.0
requests==2.32.3

pytest==8.2.2
pytest-html==4.1.1